# workspaces/bulk-import/plugins/bulk-import-backend/src/schema/openapi.yaml
...
openapi: '3.1.0'
info:
  version: '1.0'
  title: 'Bulk Import Backend'
  description: The Bulk Import Backend APIs allow users to bulk import repositories into the Backstage catalog from remote sources such as Git.
servers:
  - url: '{protocol}://{host}:{port}/{basePath}'
    variables:
      protocol:
        enum: [http, https]
        default: http
      host:
        default: localhost
      port:
        default: '7007'
      basePath:
        default: 'api/bulk-import'
paths:
...
  /organizations:
    get:
      operationId: findAllOrganizations
      summary: Fetch Organizations accessible by Backstage Github Integrations
      security:
        - BearerAuth: []
      tags: [Organization]
      parameters:
        - $ref: '#/components/parameters/pagePerIntegrationQueryParam'
        - $ref: '#/components/parameters/sizePerIntegrationQueryParam'
        - $ref: '#/components/parameters/searchQueryParam'
      responses:
        200:
          description: Organization list was fetched successfully with no errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationList'
              examples:
                multipleRepos:
                  $ref: '#/components/examples/multipleOrgs'
        500:
          description: Generic error when there are errors and no Organization is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationList'
              examples:
                repositoryListErrors:
                  $ref: '#/components/examples/orgListErrors'
  /organizations/{organizationName}/repositories:
    get:
      operationId: findRepositoriesByOrganization
      summary: Fetch Repositories in the specified GitHub organization, provided it is accessible by any of the configured GitHub Integrations.
      security:
        - BearerAuth: []
      tags: [Organization]
      parameters:
        - in: path
          name: organizationName
          description: Organization name
          required: true
          schema:
            type: string
        - in: query
          name: checkImportStatus
          description: whether to return import status. Note that this might incur a performance penalty because the import status is computed for each repository.
          schema:
            type: boolean
            default: 'false'
        - $ref: '#/components/parameters/pagePerIntegrationQueryParam'
        - $ref: '#/components/parameters/sizePerIntegrationQueryParam'
        - $ref: '#/components/parameters/searchQueryParam'
      responses:
        200:
          description: Org Repository list was fetched successfully with no errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryList'
              examples:
                multipleRepos:
                  $ref: '#/components/examples/multipleRepos'
        500:
          description: Generic error when there are errors and no Org Repository is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryList'
              examples:
                repositoryListErrors:
                  $ref: '#/components/examples/repositoryListErrors'
  /repositories:
    get:
      operationId: findAllRepositories
      summary: Fetch Organization Repositories accessible by Backstage Github Integrations
      security:
        - BearerAuth: []
      tags: [Repository]
      parameters:
        - in: query
          name: checkImportStatus
          description: whether to return import status. Note that this might incur a performance penalty because the import status is computed for each repository.
          schema:
            type: boolean
            default: 'false'
        - $ref: '#/components/parameters/pagePerIntegrationQueryParam'
        - $ref: '#/components/parameters/sizePerIntegrationQueryParam'
        - $ref: '#/components/parameters/searchQueryParam'
      responses:
        200:
          description: Repository list was fetched successfully with no errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryList'
              examples:
                multipleRepos:
                  $ref: '#/components/examples/multipleRepos'
        500:
          description: Generic error when there are errors and no repository is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryList'
              examples:
                repositoryListErrors:
                  $ref: '#/components/examples/repositoryListErrors'
  /imports:
    get:
      operationId: findAllImports
      summary: Fetch Import Jobs
      security:
        - BearerAuth: []
      tags: [Import]
      parameters:
        - $ref: '#/components/parameters/apiVersionHeaderParam'

        # The '*PerIntegration' query params are being kept for backward compatibility,
        # but the behavior depends on the API Version specified in the request headers.
        - $ref: '#/components/parameters/pagePerIntegrationQueryParamDeprecated'
        - $ref: '#/components/parameters/sizePerIntegrationQueryParamDeprecated'
        - $ref: '#/components/parameters/pageQueryParam'
        - $ref: '#/components/parameters/sizeQueryParam'
        - $ref: '#/components/parameters/sortOrderQueryParam'
        - $ref: '#/components/parameters/sortColumnQueryParam'
        - $ref: '#/components/parameters/searchQueryParam'

      responses:
        200:
          description: Import Job list was fetched successfully with no errors
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/SourceImport'
                  - $ref: '#/components/schemas/ImportJobListV2'
              examples:
                twoImports:
                  $ref: '#/components/examples/twoImports'
                multipleImportJobsV2:
                  $ref: '#/components/examples/multipleImportJobsV2'
        500:
          description: Generic error when there are errors and no Import Job is returned
          content:
            application/json:
              schema:
                oneOf:
                  - type: string
                    description: Generic error
                  - $ref: '#/components/schemas/ImportJobListV2'
              examples:
                repositoryListErrors:
                  $ref: '#/components/examples/importJobListErrors'

    post:
      operationId: createImportJobs
      summary: Submit Import Jobs
      security:
        - BearerAuth: []
      tags: [Import]
      parameters:
        - in: query
          name: dryRun
          description: whether to perform a dry-run to check if entity name collisions would occur in the catalog
          schema:
            type: boolean
            default: 'false'
      requestBody:
        description: List of Import jobs to create
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ImportRequest'
            examples:
              multipleImportRequests:
                $ref: '#/components/examples/multipleImportRequests'
      responses:
        202:
          description:
            Import Jobs request was submitted successfully to the API.
            Check the status in each item of the response body list to see their individual status.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Import'
              examples:
                twoImports:
                  $ref: '#/components/examples/twoImportJobs'
  /import/by-repo:
    get:
      operationId: findImportStatusByRepo
      summary: Get Import Status by repository
      security:
        - BearerAuth: []
      tags: [Import]
      parameters:
        - in: query
          name: repo
          description: the full URL to the repo
          schema:
            type: string
        - in: query
          name: defaultBranch
          description: the name of the default branch
          schema:
            type: string
            default: 'main'
      responses:
        200:
          description: Import Job status was determined successfully with no errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Import'
              examples:
                singleImportStatusForRepo:
                  $ref: '#/components/examples/singleImportStatusForRepo'
        500:
          description: Generic error

    delete:
      operationId: deleteImportByRepo
      summary: Delete Import by repository
      security:
        - BearerAuth: []
      tags: [Import]
      parameters:
        - in: query
          name: repo
          description: the full URL to the repo
          schema:
            type: string
        - in: query
          name: defaultBranch
          description: the name of the default branch
          schema:
            type: string
            default: 'main'
      responses:
        204:
          description: Import Job was successfully delete with no errors
        500:
          description: Generic error
  /execute-template:
    post:
      operationId: executeTemplate
      summary: Execute a scaffolder template for a list of repositories
      security:
        - BearerAuth: []
      tags:
        - Import
      requestBody:
        description: The template to execute and the repositories to run it against.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteTemplateRequest'
      responses:
        '202':
          description: Template execution request was submitted successfully to the API.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  taskIds:
                    type: array
                    items:
                      type: string
        '500':
          description: Generic error
components:
  parameters:
    apiVersionHeaderParam:
      in: header
      name: api-version
      description: |
        API version.

        ## Changelog

        ### v1 (default)
        Initial version
        #### Deprecations
        * GET /imports
          * Deprecation of 'pagePerIntegration' and 'sizePerIntegration' query parameters and introduction of new 'page' and 'size' parameters
            * 'page' takes precedence over 'pagePerIntegration' if both are passed
            * 'size' takes precedence over 'sizePerIntegration' if both are passed

        ### v2
        #### Breaking changes
        * GET /imports
          * Query parameters:
            * 'pagePerIntegration' is ignored in favor of 'page'
            * 'sizePerIntegration' is ignored in favor of 'size'
          * Response structure changed to include pagination info: instead of returning a simple list of Imports, the response is now an object containing the following fields:
            * 'imports': the list of Imports
            * 'page': the page requested
            * 'size': the requested number of Imports requested per page
            * 'totalCount': the total count of Imports
      schema:
        type: string
        enum: ['v1', 'v2']
        default: 'v1'
    pagePerIntegrationQueryParam:
      in: query
      name: pagePerIntegration
      description: the page number for each Integration
      schema:
        type: integer
        default: 1
    sizePerIntegrationQueryParam:
      in: query
      name: sizePerIntegration
      description: the number of items per Integration to return per page
      schema:
        type: integer
        default: 20
    pagePerIntegrationQueryParamDeprecated:
      in: query
      name: pagePerIntegration
      description: the page number for each Integration. **Deprecated**. Use the 'page' query parameter instead.
      deprecated: true
      schema:
        type: integer
        default: 1
    sortColumnQueryParam:
      in: query
      name: sortColumn
      description: |
        The allowed values for sorting columns:
        - repository.name: Sort by repository name.
        - repository.organization: Sort by organization URL.
        - repository.url: Sort by repository URL.
        - lastUpdate: Sort by the last time the catalog-info.yaml was updated.
        - status: Sort by the status of the catalog-info.yaml.
      required: false
      schema:
        enum:
          - repository.name
          - repository.organization
          - repository.url
          - lastUpdate
          - status
        default: repository.name
    sortOrderQueryParam:
      in: query
      name: sortOrder
      description: The order of sorting asc for ascending or desc for descending
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
        default: asc
    sizePerIntegrationQueryParamDeprecated:
      in: query
      name: sizePerIntegration
      description: the number of items per Integration to return per page. **Deprecated**. Use the 'size' query parameter instead.
      deprecated: true
      schema:
        type: integer
        default: 20
    searchQueryParam:
      in: query
      name: search
      description: returns only the items that match the search string
      schema:
        type: string
    pageQueryParam:
      # used for endpoints where pagination does not depend on the Integrations configured
      in: query
      name: page
      description: the requested page number
      schema:
        type: integer
        default: 1
    sizeQueryParam:
      # used for endpoints where pagination does not depend on the Integrations configured
      in: query
      name: size
      description: the number of items to return per page
      schema:
        type: integer
        default: 20
  schemas:
    OrganizationList:
      title: Organization List
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        errors:
          type: array
          items:
            type: string
        totalCount:
          type: integer
        pagePerIntegration:
          type: integer
        sizePerIntegration:
          type: integer
...
    RepositoryList:
      title: Repository List
      type: object
      properties:
        repositories:
          type: array
          items:
            $ref: '#/components/schemas/Repository'
        errors:
          type: array
          items:
            type: string
        totalCount:
          type: integer
        pagePerIntegration:
          type: integer
        sizePerIntegration:
          type: integer
...
    ImportJobListV2:
      title: Import Job List
      type: object
      properties:
        imports:
          type: array
          items:
            $ref: '#/components/schemas/SourceImport'
        errors:
          type: array
          items:
            type: string
        totalCount:
          type: integer
        page:
          type: integer
        size:
          type: integer
...
    SourceImport:
      title: Import Job with source it originates from
      allOf:
        - $ref: '#/components/schemas/Import'
        - type: object
          properties:
            source:
              $ref: '#/components/schemas/Source'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Backstage Permissions Framework JWT
  examples:
    multipleOrgs:
      summary: Multiple organizations
      value:
        errors: []
        organizations:
          - id: 'unique-org-id-1'
            name: 'pet-org'
            url: 'https://github.com/pet-org'
            description: 'A great Pet Org'
            totalRepoCount: 10
          - id: 'unique-org-id-2'
            name: 'org-zero'
            url: 'https://ghe.example.com/org-zero'
            totalRepoCount: 0
          - id: 'unique-id-2'
            name: 'org-one'
            url: 'https://ghe.example.com/org-one'
            description: 'Org One description'
            totalRepoCount: 1234
    orgListErrors:
      summary: Errors when listing organizations
      value:
        errors:
          - 'Github App with ID 2 failed spectacularly'
        organizations: []
    multipleRepos:
      summary: Multiple repositories
      value:
        errors: []
        repositories:
          - id: 'unique-id-1'
            name: 'pet-app'
            url: 'https://github.com/my-org/pet-app'
            organization: 'my-org'
            importStatus: 'WAIT_PR_APPROVAL'
            defaultBranch: 'main'
          - id: 'unique-id-2'
            name: 'project-zero'
            url: 'https://ghe.example.com/my-other-org/project-zero'
            organization: 'my-other-org'
            importStatus: 'PR_REJECTED'
            defaultBranch: 'dev'
          - id: 'unique-id-2'
            name: 'project-one'
            defaultBranch: 'trunk'
            url: 'https://ghe.example.com/my-other-org-2/project-one'
            organization: 'my-other-org-2'
    repositoryListErrors:
      summary: Errors when listing repositories
      value:
        errors:
          - 'Github App with ID 2 failed spectacularly'
        repositories: []
    twoImports:
      summary: Two import job requests (V1)
      value:
        - id: 'bulk-import-id-1'
          status: 'WAIT_PR_APPROVAL'
          errors: []
          approvalTool: GIT
          repository:
            name: 'pet-app'
            url: 'https://github.com/my-org/pet-app'
            organization: 'my-org'
          github:
            pullRequest:
              url: 'https://github.com/my-org/pet-app/pull/1'
              number: 1
        - id: 'bulk-import-id-2'
          status: 'PR_REJECTED'
          errors: []
          approvalTool: GIT
          repository:
            name: 'pet-app-test'
            url: 'https://github.com/my-org/pet-app-test'
            organization: 'my-org'
          github:
            pullRequest:
              url: 'https://github.com/my-org/pet-app-test/pull/10'
              number: 10
    multipleImportJobsV2:
      summary: Two import job requests (V2)
      value:
        errors: []
        page: 1
        size: 2
        totalCount: 10
        imports:
          - id: 'bulk-import-id-1'
            status: 'WAIT_PR_APPROVAL'
            errors: []
            approvalTool: GIT
            repository:
              name: 'pet-app'
              url: 'https://github.com/my-org/pet-app'
              organization: 'my-org'
            github:
              pullRequest:
                url: 'https://github.com/my-org/pet-app/pull/1'
                number: 1
          - id: 'bulk-import-id-2'
            status: 'PR_REJECTED'
            errors: []
            approvalTool: GIT
            repository:
              name: 'pet-app-test'
              url: 'https://github.com/my-org/pet-app-test'
              organization: 'my-org'
            github:
              pullRequest:
                url: 'https://github.com/my-org/pet-app-test/pull/10'
                number: 10
    importJobListErrors:
      summary: Errors when listing import jobs
      value:
        errors:
          - 'Github App with ID xyz-123 failed spectacularly'
        imports: []
...